Make PT_XLA_DEBUG max frame configurable
