+++ test/debug_tool/test_pt_xla_debug.py
@@ -8,47 +8,9 @@ import torch_xla.debug.profiler as xp
 import torch_xla.utils.utils as xu
 import torch_xla.distributed.parallel_loader as pl
 import unittest
-from typing import NamedTuple
-
-
-def check_env_flag(name, default=''):
-  return os.getenv(name, default).upper() in ['ON', '1', 'YES', 'TRUE', 'Y']
-
-
-def extract_execution_cause(lines):
-  causes = []
-  for i in range(len(lines)):
-    if 'Execution Cause' in lines[i].decode():
-      causes.append(lines[i + 1].decode())
-  return causes
-
-
-def extract_compilation_cause(lines):
-  causes = []
-  for i in range(len(lines)):
-    if 'Compilation Cause' in lines[i].decode():
-      causes.append(lines[i + 1].decode())
-  return causes
-
-
-class GraphInfo(NamedTuple):
-  hash: str
-  num_input: int
-  num_output: int
-
-
-def extract_graph_infos(lines):
-  infos = []
-  for i in range(len(lines)):
-    if 'Graph Info' in lines[i].decode():
-      hash = lines[i + 1].decode().split('Graph Hash: ')[1].strip()
-      num_input = lines[i +
-                        2].decode().split('Number of Graph Inputs:')[1].strip()
-      num_output = lines[i + 3].decode().split(
-          'Number of Graph Outputs:')[1].strip()
-      infos.append(GraphInfo(hash, int(num_input), int(num_output)))
-
-  return infos
+from extract_debug_helper import (check_env_flag, extract_execution_cause,
+                                  extract_compilation_cause, GraphInfo,
+                                  extract_graph_infos)
 
 
 class PtXLADebugTest(unittest.TestCase):
